UPDATE OVERHAUL
Began working on inputs and using ASCII codes with chr() functions, Learned and used some embedded Python functions to check for whether an ASCII value represents an alphanumeric character. Started working on inputs and how to ask for whether the user would like special characters and how they should be placed within the output password.

TO DO:
Finish up the former. Conjoin the contents of the passward_chars list output to display as a string without spaces or commas.